Issue #3003810 by bramtenhove, Kingdutch: Ensure proper behaviour for unpublished comment reply links
The reply link was displaying incorrectly. In the proper logic for
this code the link is set to an inactive display when a comment is
unpublished. Then the normal logic runs to determine whether to show
the link (or unpublished placeholder) at all.

This ensures that the word "Reply" is only shown in the correct
places and ensures that a comment's structure is always the same
whether it's published or not. This is the desired UX that was
determined by Xinyu.